services/vanna/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
psycopg[binary]==3.1.18
pydantic==2.7.1